#23 - Add Registrator in cpuAV.h
